move comment box stuff to its own folder
